+++ scenarios/__init__.py
@@ -1,0 +1,1 @@
+#from .common.infrastructure_deployment import provision_terraform_templates
